requirements.txt
torch
torchvision
ftfy
regex
Pillow
numpy
flask
faiss-cpu
git+https://github.com/openai/CLIP.git
